Add files via upload
gannsquare is an interesting concept, where the support and resistance lines are predicted using the Fibonacci series. I have the module readily available in gannsquare.py. 
The idea is to use ema/trend.gann strategies to find the right time to enter a trade. 
This is an extended version of trading bot, still very raw , will release the final version in few days